Moved WbPosixSharedMemory class into a file
